pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "nwbinspector"
version = "0.6.0"
description = "Tool to inspect NWB files for best practices compliance."
readme = "README.md"
authors = [
    {name = "Cody Baker"},
    {name = "Steph Prince"},
    {name = "Szonja Weigl"},
    {name = "Heberto Mayorquin"},
    {name = "Paul Adkisson"},
    {name = "Luiz Tauffer"},
    {name = "Ben Dichter",  email = "ben.dichter@catalystneuro.com"}
]
urls = { "Homepage" = "https://github.com/NeurodataWithoutBorders/nwbinspector" }
license = {file = "license.txt"}
keywords = ["nwb"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "License :: OSI Approved :: BSD License",
]
requires-python = ">=3.9"
dependencies = [
    "pynwb>=2.8",   # NWB Inspector should always be used with most recent minor versions of PyNWB
    "hdmf-zarr",
    "fsspec",
    "s3fs",
    "requests",
    "aiohttp",
    "PyYAML",
    "jsonschema",
    "packaging",
    "natsort",
    "click",
    "tqdm",
    "isodate",
    "numpy>=1.22.0,<2.0.0", # TODO: add compatibility for 2.0 - HDMF Zarr and others also still have some troubles
    "h5py<3.12.0", # TODO: remove pin when https://github.com/h5py/h5py/issues/2505 is fixed and released
]

[project.optional-dependencies]
dandi = [
    "dandi",
    "remfile",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
nwbinspector = "nwbinspector._nwbinspector_cli:_nwbinspector_cli"



[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.toml
  |\.rst
  |\.md
  |\.yml
  |\.txt
  |\.sh
  |\.git
  |\.ini
  |\.hg
  |\.mypy_cache
  |\.tox
  |\.venv
  |build
  |dist
)/
'''



[tool.ruff]
exclude = ["docs/*"]

[tool.ruff.lint]
# TODO: eventually, expand to other 'D', and other 'F' linting
select = [
    "F401",
    "F541",
    "I",
    "ANN001",
    "ANN201",
    "ANN202",
    "ANN205",
    "ANN206",
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"src/nwbinspector/__init__.py" = ["I"]  # Must maintain explicit order for imports for check registration
"src/nwbinspector/utils/__init__.py" = ["F401", "I"]  # Several items planned for long-term deprecation

# Temporarily keeping all exposure below for back-compatibility
# Deprecation scheduled for 9/15/2024
"src/nwbinspector/nwbinspector/__init__.py" = ["F401", "I"]
"src/nwbinspector/inspector_tools/__init__.py" = ["F401", "I"]
"src/nwbinspector/version/__init__.py" = ["F401", "I"]
"src/nwbinspector/register_checks/__init__.py" = ["F401", "I"]

[tool.ruff.lint.isort]
relative-imports-order = "closest-to-furthest"
known-first-party = ["nwbinspector"]



[tool.codespell]
skip = '.git*,*.pdf,*.css'
check-hidden = true
ignore-words-list = 'assertin'
